LW altered structure of beeps and pauses
Altered structure of beeps and pauses, and added in additional fwrite
code to force all LEDs off between intervals (the arduino doesn’t seem
to do a good job of this!?).  The sound is still a bib off, but more so
in earlier trials, this needs to be looked at.