LW updated the calibration data that gets imported
The ‘4LEDS’ version of the code was still using the calibration .mat
file that had all 5 LEDs, and then was using the first 4 LED
calibrations (which included the non-working one, and missed out a
working one).  A .mat data file excluding the broken LED was saved and
is now used in the code.